+++ check-update-all.sh
@@ -12,6 +12,10 @@ cargo clippy --all --all-targets -- -D warnings
 cargo nextest run -p rustemo-compiler
 cargo install --path rustemo-compiler --debug
 
+# README Examples
+rcomp docs/src/readme_example/src/textlr/calclr.rustemo
+rcomp --parser-algo glr docs/src/readme_example/src/textglr/calc.rustemo
+
 cd docs/src/tutorials/calculator/
 for i in {1..5}; do
     rcomp calculator$i/src/calculator.rustemo;